+++ tests/jsonql.test.js
@@ -51,3 +51,59 @@ test('accept json as query', () => {
         + '&or[_embedded.foo.bar.name]=Simone'
     );
 });
+
+test('query al fields with or using or() alias', () => {
+    var jsonQuery = {
+        'relation.nick': '@sensorario',
+        'foo.bar.name': 'Simone'
+    };
+    var queryQl = new QueryQl()
+    queryQl.or(jsonQuery);
+    expect(queryQl.getQueryString()).toEqual(
+        'rel=relation,foo,bar'
+        + '&or[_embedded.relation.nick]=@sensorario'
+        + '&or[_embedded.foo.bar.name]=Simone'
+    );
+});
+
+test('query all fields with and using and() alias', () => {
+    var jsonQuery = {
+        'relation.nick': '@sensorario',
+        'foo.bar.name': 'Simone'
+    };
+    var queryQl = new QueryQl()
+    queryQl.and(jsonQuery);
+    expect(queryQl.getQueryString()).toEqual(
+        'rel=relation,foo,bar'
+        + '&and[_embedded.relation.nick]=@sensorario'
+        + '&and[_embedded.foo.bar.name]=Simone'
+    );
+});
+
+test('query all fields with filtering using andFiltering() alias', () => {
+    var jsonQuery = {
+        'relation.nick': '@sensorario',
+        'foo.bar.name': 'Simone'
+    };
+    var queryQl = new QueryQl()
+    queryQl.andFiltering(jsonQuery);
+    expect(queryQl.getQueryString()).toEqual(
+        'rel=relation,foo,bar'
+        + '&filtering[_embedded.relation.nick]=@sensorario'
+        + '&filtering[_embedded.foo.bar.name]=Simone'
+    );
+});
+
+test('query all fields with filtering_or using orFiltering() alias', () => {
+    var jsonQuery = {
+        'relation.nick': '@sensorario',
+        'foo.bar.name': 'Simone'
+    };
+    var queryQl = new QueryQl()
+    queryQl.orFiltering(jsonQuery);
+    expect(queryQl.getQueryString()).toEqual(
+        'rel=relation,foo,bar'
+        + '&filtering_or[_embedded.relation.nick]=@sensorario'
+        + '&filtering_or[_embedded.foo.bar.name]=Simone'
+    );
+});